Rename useCs to useSubscription.
